Utilities & screenBuffer
Datei Utilities für generell nützliche Funktionen

screenBuffer ais main in eigene hpp & cpp Dateien ausgelagert
