+++ CanLogger/include/utilities.hpp
@@ -1,0 +1,7 @@
+#ifndef UTILITIES_HPP
+#define UTILITIES_HPP
+
+void processCanMessage(void);
+void AddZerosToString(String& s,uint64_t const val,uint64_t const maxVal, uint8_t const numberSys);
+
+#endif //UTILITIES_HPP
