+++ CanLogger/src/main.cpp
@@ -6,23 +6,13 @@
 #include "can/Canmsg.hpp"
 #include "serial/SerialCommunication.hpp"
 #include "buttons/Encoder.hpp"
+#include "display/screenBuffer.hpp"
+#include "utilities.hpp"
 
 Adafruit_ILI9341 display = Adafruit_ILI9341{PC9, PA8, PA10, PB5, PC8}; // use Software Serial
 //Adafruit_ILI9341 display = Adafruit_ILI9341{PC9, PA8, PC8}; // use Hardware Serial
 
 using namespace utilities; // für scom
-
-//start Definitionen fuer Displaybuffer
-void printScreenBufferSerial(void);
-void processCanMessage(void);
-void insertMessageHere(Canmsg const& msg, int pos);
-void makeBufferVisible(void);
-constexpr int SCREEN_BUFFER_SIZE = 10;
-static int screenBufferFillLevel = 0;
-static int screenBufferUserViewFillLevel = 0;
-static Canmsg* screenBuffer = nullptr;
-static Canmsg* screenBufferUserView = nullptr;
-//end Definitionen fuer Displaybuffer
 
 void setup() {
   HAL_Init();
@@ -38,11 +28,9 @@ void setup() {
   {
     while(1){}
   }
-  screenBuffer = new Canmsg[SCREEN_BUFFER_SIZE];
-  screenBufferUserView = new Canmsg[SCREEN_BUFFER_SIZE];
-  screenBufferFillLevel = 0;
-  screenBufferUserViewFillLevel = 0;
   
+  screenBufferInit();
+
   display.begin();
   display.fillScreen(ILI9341_BLACK);
   display.setTextSize(3);
@@ -59,155 +47,8 @@ void loop() {
     processCanMessage();
   }
   makeBufferVisible();
-  printScreenBufferSerial();
+  printScreenBufferUserViewSerial();
   delay(5000);
-}
-
-/* 
-    function that prints the content of the screenBuffer via the Serial Port
-*/
-void printScreenBufferSerial(void)
-{
-  Serial.print("Ausgabebuffer Fuellstand: ");
-  Serial.print(screenBufferUserViewFillLevel);
-  Serial.print("/");
-  Serial.println(SCREEN_BUFFER_SIZE);
-  for(int i=0; i<screenBufferUserViewFillLevel; i++)
-  {
-    Serial.print("Nachricht ");
-    if((i+1)<10)
-    {
-      Serial.print(" ");
-    }
-    Serial.print(i+1);
-    Serial.print("/");
-    Serial.print(screenBufferUserViewFillLevel);
-    Serial.print(": ");
-    Serial.println(static_cast<String>(screenBufferUserView[i]));
-  }
-  Serial.println("\r\n");
-}
-
-/* 
-    inserts the surpassed message at the surpassed position
-    Input:  msg - reference to the message that should be inserted
-            pos - position the message should be Placed
-                  note that if the buffer is allready filled, the last message will be discarded
-                  possible values: 0 - screenBufferFillLevel
-*/
-void insertMessageHere(Canmsg const& msg, int pos)
-{
-  if((pos < SCREEN_BUFFER_SIZE) && (pos <= screenBufferFillLevel))
-  {
-    if(screenBufferFillLevel >= SCREEN_BUFFER_SIZE)
-    {
-      screenBufferFillLevel = SCREEN_BUFFER_SIZE-1;
-    }
-    screenBuffer[screenBufferFillLevel] = msg;
-    for(int i=screenBufferFillLevel; i>pos; i--)
-    {
-      Canmsg temp = std::move(screenBuffer[i]); 
-      screenBuffer[i] = std::move(screenBuffer[i-1]);
-      screenBuffer[i-1] = std::move(temp);
-    }
-    if(screenBufferFillLevel < SCREEN_BUFFER_SIZE)
-    {
-      screenBufferFillLevel++;
-    }
-  }
-}
-
-/* 
-    function that processes all the pending CAN-messages in the Canmsg_bufferCanRecMessages
-*/
-void processCanMessage(void)
-{
-  Canmsg_bufferCanRecPointer--;
-  Canmsg curMsg = std::move(Canmsg_bufferCanRecMessages[Canmsg_bufferCanRecPointer]);
-  
-  //screenBuffer:
-  bool messageInserted = false;
-  for(int i=0; i<screenBufferFillLevel && !messageInserted; i++)
-  {
-    if(!curMsg.GetIsExtIdentifier())
-    {
-      if(!screenBuffer[i].GetIsExtIdentifier())
-      {
-        if(curMsg.GetStdIdentifier() == screenBuffer[i].GetStdIdentifier())
-        {
-          screenBuffer[i] = curMsg;
-          messageInserted = true;
-        }
-        else if(curMsg.GetStdIdentifier() < screenBuffer[i].GetStdIdentifier())
-        {
-          insertMessageHere(curMsg, i);
-          messageInserted = true;
-        }
-      }
-      else
-      {
-        insertMessageHere(curMsg, i);
-        messageInserted = true;
-      }
-    }
-    else
-    {
-      if(screenBuffer[i].GetIsExtIdentifier())
-      {
-        if(curMsg.GetStdIdentifier() == screenBuffer[i].GetStdIdentifier())
-        {
-          if(curMsg.GetExtIdentifier() == screenBuffer[i].GetExtIdentifier())
-          {
-            screenBuffer[i] = curMsg;
-            messageInserted = true;
-          }
-          else if(curMsg.GetExtIdentifier() < screenBuffer[i].GetExtIdentifier())
-          {
-            insertMessageHere(curMsg, i);
-            messageInserted = true;
-          }
-        }
-        else if(curMsg.GetStdIdentifier() < screenBuffer[i].GetStdIdentifier())
-        {
-          insertMessageHere(curMsg, i);
-          messageInserted = true;
-        }
-      }
-    }
-  }
-  if((screenBufferFillLevel < SCREEN_BUFFER_SIZE) && !messageInserted)
-  {
-    screenBuffer[screenBufferFillLevel] = curMsg;
-    screenBufferFillLevel++;
-  }
-
-  //SD-Card:
-
-  //loopback:
-  /*
-  Canmsg_bufferCanRecMessages[Canmsg_bufferCanRecPointer].Send();
-  */
-
-  //Serial:
-  /*
-  Serial.print("Empfangene Nachricht: ");
-  Serial.println(static_cast<String>(Canmsg_bufferCanRecMessages[Canmsg_bufferCanRecPointer]));
-  */
-}
-
-/* 
-    copies the messages from the Backendbuffer to the Frontendbuffer
-*/
-void makeBufferVisible(void)
-{
-  for(int i=0; i<screenBufferFillLevel; i++)
-  {
-    if(screenBufferUserView[i] != screenBuffer[i])
-    {
-      screenBufferUserView[i] = screenBuffer[i];
-    }
-  }
-  screenBufferUserViewFillLevel = screenBufferFillLevel;
 }
 
 void serialEvent() {
